+++ packages/design-system/jest.config.js
@@ -1,5 +1,0 @@
-const talendScripts = require('@talend/scripts-config-jest/jest.config');
-
-module.exports = {
-	...talendScripts,
-};