+++ packages/design-system/src/components/Form/Field/Input/Input.Copy.tsx
@@ -1,40 +1,59 @@
 import React from 'react';
 import { useCopyToClipboard } from 'react-use';
+import { useTranslation } from 'react-i18next';
 import Button from '../../../Button';
 import FieldGroup from '../../FieldGroup';
 import Text from './Input.Text';
 import { InputProps } from './Input';
 
 const InputCopy = React.forwardRef(
-	({ label, disabled, readOnly, ...rest }: InputProps, ref: React.Ref<HTMLInputElement | null>) => {
-		const [state, copyToClipboard] = useCopyToClipboard();
+	({ label, disabled, readOnly, value, defaultValue,...rest }: InputProps, ref: React.Ref<HTMLInputElement | null>) => {
+		const [{value: copiedValue, error: copyError}, copyToClipboard] = useCopyToClipboard();
 		const inputRef = React.useRef<HTMLInputElement | null>(null);
+		const { t } = useTranslation();
+		const inputValue = value || defaultValue;
 
 		React.useImperativeHandle(ref, () => inputRef.current);
+
+		const getDescriptionMessage = () => {
+			if(copyError) {
+				return copyError.message;
+			} else if (copiedValue && copiedValue === inputValue) {
+				return t('FORM_COPY_COPIED_TO_CLIPBOARD', 'Copied to clipboard')
+			} else {
+				return '';
+			}
+		};
 
 		return (
 			<FieldGroup
 				label={label}
 				suffix={
-					!readOnly && (
+					!readOnly ? (
 						<Button.Icon
 							icon="talend-files-o"
 							onClick={() => copyToClipboard(inputRef.current?.value || '')}
 							disabled={disabled}
 						>
-							Copy to clipboard
+							{t('FORM_COPY_COPY_TO_CLIPBOARD', 'Copy to clipboard')}
 						</Button.Icon>
-					)
+					) : undefined
 				}
 				readOnly={!disabled}
 				disabled={!!disabled}
-				hasError={!!state.error}
-				hasSuccess={!!state.value}
-				description={state.error ? state.error.message : state.value && 'Copied to clipboard'}
+				hasError={!!copyError}
+				hasSuccess={!!copiedValue}
+				description={getDescriptionMessage()}
 			>
 				{/*
 				// @ts-ignore */}
-				<Text {...rest} label="Copy to clipboard" ref={inputRef} />
+				<Text
+					{...rest}
+					value={value}
+					defaultValue={defaultValue}
+					label={t('FORM_COPY_COPY_TO_CLIPBOARD', 'Copy to clipboard')}
+					ref={inputRef}
+				/>
 			</FieldGroup>
 		);
 	},