+++ packages/design-system/src/components/Stack/index.ts
@@ -1,0 +1,3 @@
+export { StackHorizontal } from './StackHorizontal';
+export { StackVertical } from './StackVertical';
+export { StackItem } from './StackItem';