+++ packages/design-system/src/components/Stepper/Step/variations/Step.error.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
 import Step from '../Step';
-import tokens from '../../../../tokens';
 
 const StepError = styled(Step).attrs({
 	icon: 'talend-cross-circle',