+++ packages/design-system/src/components/ThemeProvider/ThemeSwitcher/ThemeSwitcher.tsx
@@ -1,4 +1,5 @@
 import React, { useContext, useState } from 'react';
+import { useTranslation } from 'react-i18next';
 
 import Toggle from '../../Toggle';
 import ThemeContext from '../ThemeContext';
@@ -8,20 +9,21 @@ import { dark, light } from '../../../themes';
 const ThemeSwitcher = () => {
 	const { switchTheme, theme } = useContext(ThemeContext);
 	const [hasDarkMode, setDarkMode] = useState(false);
+	const { t } = useTranslation();
 
 	React.useEffect(() => {
 		setDarkMode(theme === dark);
 	}, [theme]);
 
-	function toggle() {
+	const toggle = () => {
 		if (switchTheme) {
 			switchTheme(hasDarkMode ? light : dark);
 		}
-	}
+	};
 
 	return (
 		<Toggle icon={hasDarkMode ? 'talend-eye-slash' : 'talend-eye'} onChange={toggle}>
-			Toggle dark mode
+			{t('THEME_TOGGLE_DARK_MODE', 'Toggle dark mode')}
 		</Toggle>
 	);
 };