+++ packages/design-system/src/pages/Home.tsx
@@ -5,6 +5,8 @@ import Menu from './blocks/Menu';
 
 import Template from '../templates';
 
-const Home = props => <Template.List header={<Header />} nav={<Menu />} {...props} />;
+const Home = (props: React.PropsWithChildren<any>) => (
+	<Template.List header={<Header />} nav={<Menu />} {...props} />
+);
 
 export default Home;