+++ packages/design-system/src/pages/index.ts
@@ -1,9 +1,4 @@
 import Login from './Login';
 import Home from './Home';
 
-const Page = {};
-
-Page.Login = Login;
-Page.Home = Home;
-
-export default Page;
+export { Login, Home };