+++ packages/design-system/src/themes/index.ts
@@ -1,9 +1,4 @@
 import light from './light.theme';
 import dark from './dark.theme';
 
-import lightDictionary from './light/dictionary';
-import darkDictionary from './dark/dictionary';
-
-import './themes.scss';
-
-export { light, lightDictionary, dark, darkDictionary };
+export { light, dark };